agentic-ai-api-react/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
chromadb
google-generativeai
pypdf
